requirements.txt
Flask
python-telegram-bot[job-queue]
requests
pandas

